+++ project.log
@@ -1,0 +1,6 @@
+INFO:root:Number of male students: 75
+INFO:root:Number of female students: 48
+INFO:root:Students with special characters in names: ["Akong'o, Shanelle Hilda Anyango", "Ang'edu, Jude Ocomil", "Ng'ang'a, Ian Mutuge", "Orony, Charis Ng'wono", "Waweru, Monika Ng'endo", "Mbugua, Nathan Ng'ethe", "Nyang'or, Olive Menorah"]
+INFO:root:Number of male students: 75
+INFO:root:Number of female students: 48
+INFO:root:Students with special characters in names: ["Akong'o, Shanelle Hilda Anyango", "Ang'edu, Jude Ocomil", "Ng'ang'a, Ian Mutuge", "Orony, Charis Ng'wono", "Waweru, Monika Ng'endo", "Mbugua, Nathan Ng'ethe", "Nyang'or, Olive Menorah"]
